+++ routes/review.js
@@ -1,10 +1,12 @@
 const reviewRoute = require('express').Router();
 const reviewCtrl = require('../controllers').controller.Review;
+const authenticate = require('../middleware/authenticate');
+const role = require('../middleware/role');
 
-reviewRoute.get('/s', reviewCtrl.getMany);
-reviewRoute.get('/:id', reviewCtrl.getById);
-reviewRoute.post('/', reviewCtrl.createData);
-reviewRoute.put('/:id', reviewCtrl.updateById);
-reviewRoute.delete('/:id', reviewCtrl.deleteById);
+reviewRoute.get('/s', authenticate.authenticateJWT, role.checkRole(["Admin", "Customer"], { read: true }), reviewCtrl.getMany);
+reviewRoute.get('/:id', authenticate.authenticateJWT, role.checkRole(["Admin"], { read: true }), reviewCtrl.getById);
+reviewRoute.post('/', authenticate.authenticateJWT, role.checkRole(["Admin", "Customer"], { create: true }), reviewCtrl.createData);
+reviewRoute.put('/:id', authenticate.authenticateJWT, role.checkRole(["Admin", "Customer"], { update: true }), reviewCtrl.updateById);
+reviewRoute.delete('/:id', authenticate.authenticateJWT, role.checkRole(["Admin", "Customer"], { delete: true }), reviewCtrl.deleteById);
 
 module.exports = reviewRoute;
