+++ routes/tour.js
@@ -1,10 +1,12 @@
 const tourRoute = require('express').Router();
 const tourCtrl = require('../controllers').controller.Tour;
+const authenticate = require('../middleware/authenticate');
+const role = require('../middleware/role');
 
-tourRoute.get('/s', tourCtrl.getMany);
-tourRoute.get('/:id', tourCtrl.getById);
-tourRoute.post('/', tourCtrl.createData);
-tourRoute.put('/:id', tourCtrl.updateById);
-tourRoute.delete('/:id', tourCtrl.deleteById);
+tourRoute.get('/s', role.checkRoleQuery(["code"], ["Admin"], { read: true }), tourCtrl.getMany);
+tourRoute.get('/:id', authenticate.authenticateJWT, role.checkRole(["Admin"], { read: true }), tourCtrl.getById);
+tourRoute.post('/', authenticate.authenticateJWT, role.checkRole(["Admin"], { create: true }), tourCtrl.createData);
+tourRoute.put('/:id', authenticate.authenticateJWT, role.checkRole(["Admin"], { update: true }), tourCtrl.updateById);
+tourRoute.delete('/:id', authenticate.authenticateJWT, role.checkRole(["Admin"], { delete: true }), tourCtrl.deleteById);
 
 module.exports = tourRoute;
